Move utils to top level
