+++ erftools/postprocessing/__init__.py
@@ -1,7 +1,7 @@
 from .surface import SurfaceHistory
 
 from .profiles import AveragedProfiles
-from .utils import stagger_profile, destagger_profile
+from .helpers import stagger_profile, destagger_profile
 
 from .pltfiles import Plotfile
